[modify] #9 mofify view for PC
