+++ .config/configstore/update-notifier-npm.json
@@ -1,0 +1,4 @@
+{
+	"optOut": false,
+	"lastUpdateCheck": 1553012306506
+}
