+++ .gitignore
@@ -33,3 +33,5 @@ yarn-debug.log*
 
 /.env
 /public/entry
+/app/frontend/stylesheets/scss
+/app/frontend/stylesheets/webfonts
