+++ .npm/_logs/2019-03-19T16_18_26_623Z-debug.log
@@ -1,0 +1,23 @@
+0 info it worked if it ends with ok
+1 verbose cli [ '/usr/bin/node', '/usr/bin/npm', 'run', 'build' ]
+2 info using npm@6.4.1
+3 info using node@v8.15.1
+4 verbose config Skipping project config: /app/.npmrc. (matches userconfig)
+5 verbose stack Error: missing script: build
+5 verbose stack     at run (/usr/lib/node_modules/npm/lib/run-script.js:155:19)
+5 verbose stack     at /usr/lib/node_modules/npm/lib/run-script.js:63:5
+5 verbose stack     at /usr/lib/node_modules/npm/node_modules/read-package-json/read-json.js:115:5
+5 verbose stack     at /usr/lib/node_modules/npm/node_modules/read-package-json/read-json.js:418:5
+5 verbose stack     at checkBinReferences_ (/usr/lib/node_modules/npm/node_modules/read-package-json/read-json.js:373:45)
+5 verbose stack     at final (/usr/lib/node_modules/npm/node_modules/read-package-json/read-json.js:416:3)
+5 verbose stack     at then (/usr/lib/node_modules/npm/node_modules/read-package-json/read-json.js:160:5)
+5 verbose stack     at /usr/lib/node_modules/npm/node_modules/read-package-json/read-json.js:364:12
+5 verbose stack     at /usr/lib/node_modules/npm/node_modules/graceful-fs/graceful-fs.js:78:16
+5 verbose stack     at FSReqWrap.readFileAfterClose [as oncomplete] (fs.js:511:3)
+6 verbose cwd /app
+7 verbose Linux 4.9.125-linuxkit
+8 verbose argv "/usr/bin/node" "/usr/bin/npm" "run" "build"
+9 verbose node v8.15.1
+10 verbose npm  v6.4.1
+11 error missing script: build
+12 verbose exit [ 1, true ]
