+++ app/frontend/javascripts/application.js
@@ -1,3 +1,3 @@
 import 'jquery-ujs/src/rails';
-require('./ripple.js');
 window.$ = window.jQuery = require('jquery');
+window.Main = require('./main');
