+++ app/frontend/stylesheets/application.scss
@@ -1,2 +1,2 @@
-@import 'reset';
-@import 'main';
+@import './reset';
+@import './main';
