+++ app/frontend/stylesheets/font_awesome/scss/_fixed-width.scss
@@ -1,0 +1,6 @@
+// Fixed Width Icons
+// -------------------------
+.#{$fa-css-prefix}-fw {
+  text-align: center;
+  width: $fa-fw-width;
+}
