+++ app/frontend/stylesheets/font_awesome/scss/v4-shims.scss
@@ -1,0 +1,2 @@
+@import 'variables';
+@import 'shims';
