+++ app/frontend/stylesheets/main.scss
@@ -1,4 +1,10 @@
 @import url('https://fonts.googleapis.com/css?family=Noto+Sans+JP:400,700&subset=japanese');
+$fa-font-path: "./font_awesome/webfonts";
+@import "./font_awesome/scss/fontawesome.scss";
+@import "./font_awesome/scss/brands.scss";
+@import "./font_awesome/scss/solid.scss";
+
+//common
 $pink:#ef6d6e;
 $sakura:#FFB2AE;
 $ume:#fcecea;
@@ -10,17 +16,8 @@ body{
     color:#fff;
     font-family: 'Noto Sans JP', 'ヒラギノ角ゴシック', 'メイリオ' ,sans-serif;
     font-size: 1em;
-    height:100vh;
+    height:100vh　- 4em;
     width:100vw;
-}
-.main{
-    height:100%;
-    margin:0 30%;
-    text-align: center;
-    width:40%;
-}
-.container{
-    padding:10% 0;
 }
 
 .btn{
@@ -47,36 +44,14 @@ body{
 .btn-white{
     background-color:$white;
     color:$pink;
+    &:hover{
+        background-color: $pink;
+        color:$white;
+    }
 }
 
-.logo{
-    border-radius: 50%;
-    height:7em;
-    margin:2em;
-    width:7em;
-}
-.card{
-    border-radius: 0.2em;
-    margin:1em 0;
-    padding:1em 3em;
-}
-.card-white{
-    background-color: $white;
-    color:$pink;
-}
-li{
-    margin:1em 0;
-}
 .shadow{
     box-shadow: 0 2px 2px 0 rgba(0,0,0,0.14), 0 1px 5px 0 rgba(0,0,0,0.12), 0 3px 1px -2px rgba(0,0,0,0.2);
-}
-#content_original_text{
-    background-color: $white;
-    font-size: 1em;
-    height:10em;
-    margin:2em 0;
-    padding: 1em;
-    width:100%;
 }
 
 .ripple{
@@ -106,4 +81,133 @@ li{
         opacity: 0;
         transform: scale(2.0);
     }
+}
+
+//main
+.main{
+    height:100%;
+    margin:0 30%;
+    text-align: center;
+    width:40%;
+    .container{
+        padding:10% 0;
+        .logo{
+            border-radius: 50%;
+            height:7em;
+            margin:2em;
+            width:7em;
+        }
+        .card{
+            border-radius: 0.2em;
+            margin:1em 0;
+            padding:1em 3em;
+        }
+        .card-white{
+            background-color: $white;
+            color:$pink;
+            &:hover{
+                background-color: $sakura;
+                color:$white;
+            }
+            .hide{
+                display:none;
+            }
+        }
+        li{
+            margin:1em 0;
+        }
+        #content_original_text{
+            background-color: $white;
+            font-size: 1em;
+            height:10em;
+            margin:2em 0;
+            padding: 1em;
+            width:100%;
+        }
+        .flash{
+            background-color: $sakura;
+            border-radius: 0.2em;
+            color:$white;
+            margin-bottom:3em;
+            text-align: left;
+            padding:2em;
+            position: relative;
+            width:100%;
+        }
+        .flash-success{
+            @extend %fa-icon;
+            @extend .fas;
+            &:after{
+                content:fa-content($fa-var-check);
+                right:2em;
+                position: absolute;
+            }
+        }
+        .flash-danger{
+            @extend %fa-icon;
+            @extend .fas;
+            &:after{
+                content:fa-content($fa-var-times);
+                right:2em;
+                position: absolute;
+            }
+        }
+    }
+}
+
+//header
+header{
+    background-color:$pink;
+    display:flex;
+    height:4em;
+    justify-content: space-between;
+    .header-left{
+        align-items: center;
+        color: $white;
+        display: flex;
+        .header-logo{
+            border-radius: 50%;
+            height:4em;
+            width: 4em;
+        }
+    }
+    .header-right{
+        @extend %fa-icon;
+        @extend .fas;
+        align-items: center;
+        display: flex;
+        .header-item{
+            font-size: 1.2em;
+            margin-right: 2em;
+            a{
+                color:$white;
+                &:hover{
+                    color:$sakura;
+                }
+            }
+        }
+        .icon-home{
+            &:before {
+               content: fa-content($fa-var-home);
+            }
+        }
+        .icon-new{
+            &:before{
+                content: fa-content($fa-var-pen)
+            }
+        }
+        .icon-index{
+            &:before{
+                content: fa-content($fa-var-comments)
+            }
+        }
+        .icon-twitter{
+            @extend %fa-icon;
+            @extend .fab;
+            &:before{
+                content: fa-content($fa-var-twitter)
+            }
+        }
+    }
+    
 }
