+++ app/views/contents/index.html.erb
@@ -3,6 +3,7 @@
 <% end %>
 <% @contents.each do |content| %>
     <div class="card card-white shadow">
-        <%= "#{content.original_text} |  #{content.converted_text}" %>
+        <span class="original"><%= content.original_text %></span>
+        <span class="converted hide"><%= content.converted_text %></span>
     </div>
 <% end %>
