+++ app/views/contents/success.html.erb
@@ -1,13 +1,12 @@
 <ul>
-    <li><%=link_to "ツイッター　ミル", "https://twitter.com/genshijin_bot", class: "btn btn-white shadow" %></li>
+    <li><%=link_to "ツイッター　ミル", "https://twitter.com/genshijin_bot", class: "btn btn-white shadow ripple" %></li>
     <li>
         <% if request.referrer.include?('new') %>
-            <%= link_to "モウイッカイ",new_content_path, class: "btn btn-pink shadow" %>
+            <%= link_to "モウイッカイ",new_content_path, class: "btn btn-white shadow ripple" %>
         <% else %>
             <%= form_for @content, url: tweet_path do |f| %>
-                <%= f.submit "モウイッカイ", class: "btn btn-pink shadow" %>
+                <%= f.submit "モウイッカイ", class: "btn btn-white shadow ripple" %>
             <% end %>
         <% end %>
     </li>
-    <li><%=link_to "イエ　カエル",root_path, class: "btn btn-sakura shadow" %></li>
 </ul>
