+++ config/webpacker.yml
@@ -26,6 +26,7 @@ default: &default
     - .gif
     - .jpeg
     - .jpg
+    
 
 development:
   <<: *default
